// types.ts
export interface NavItem {
  label: string;
  href: string;
}

export interface ServiceItem {
  id: string;
  no: string;
  name: string;
  details: string;
  price: string;
  image: string;
}

export interface Testimonial {
  id: string;
  image: string;
  quote: string;
  names: string;
  date: string;
}